+++ app/src/main/java/com/inky/fitnesscalendar/ui/views/RecordActivityDialog.kt
@@ -1,6 +1,5 @@
 package com.inky.fitnesscalendar.ui.views
 
-import android.content.Context
 import android.os.Parcelable
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxWidth
@@ -14,7 +13,6 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import com.inky.fitnesscalendar.R
@@ -37,9 +35,8 @@ fun RecordActivity(
     localizationRepository: LocalizationRepository,
     onNavigateBack: () -> Unit
 ) {
-    val context = LocalContext.current
     var state by rememberSaveable {
-        mutableStateOf(RecordActivityState.fromPrediction(context))
+        mutableStateOf(RecordActivityState.fromPrediction())
     }
 
     val typeRows = localDatabaseValues.current.activityTypeRows
@@ -149,10 +146,10 @@ data class RecordActivityState(
     }
 
     companion object {
-        fun fromPrediction(context: Context) = RecordActivityState(
+        fun fromPrediction() = RecordActivityState(
             activitySelectorState = ActivitySelectorState.fromPrediction(
-                context,
-                requireTypeHasDuration = true
+                requireTypeHasDuration = true,
+                selectedActivityType = null
             ),
             customStart = null
         )